requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
cohere==5.13.6
distro==1.9.0
fastavro==1.10.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
jiter==0.8.2
openai==1.59.3
packaging==24.2
parameterized==0.9.0
PyAudio==0.2.14
pyttsx3==2.90
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
whisper==1.1.10
wikipedia==1.4.0
